:memo: Update pydoc for CLI fixup
* Add explanation for CLI fixup
* Move fixup functions out of cli/__init__.py
